ethdev: fix enabling RSS behavior inconsistent
The RTE_ETH_MQ_RX_RSS_FLAG flag is a switch to enable RSS. If the flag is
not set in dev_configure, RSS will be not configured and enabled. However,
RSS hash and reta can still be configured by ethdev ops to enable RSS if
the flag isn't set. The behavior is inconsistent.

Fixes: 99a2dd955fba ("lib: remove librte_ prefix from directory names")
Cc: stable@dpdk.org

Signed-off-by: Huisong Li <lihuisong@huawei.com>
Signed-off-by: Min Hu (Connor) <humin29@huawei.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>