kni: optimize alloc queue release
the kni alloc queue is filled with physical addresses of mbufs for kernel
consumption. Any unused mbufs in the alloc queue are freed during
shutdown sequence in rte_kni_release.

In the current existing implementation, for freeing one entry of alloc queue
all the objects of the mempool are traversed. This process is repeated
for all the objects of the alloc queue which takes lot of cpu cycles.

Instead of using mempool object iteration method,
use ``rte_mem_iova2virt()`` api to get the virtual
address for the physical addresses of alloc_q objects.
This sppeds up the freeing process.

Signed-off-by: Naga Harish K S V <s.v.naga.harish.k@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>