net/virtio: unmap PCI device in secondary process
In multi-process, the secondary process will remap PCI during
initialization, but the mapping is not removed in the uninit path,
the device is not closed, and the device busy error will be reported
when the device is hotplugged.

This patch unmap PCI device at secondary process uninitialization
based on virtio_rempa_pci().

Fixes: 36a7a2e7a53 ("net/virtio: move PCI device init in dedicated file")

Signed-off-by: Yuan Wang <yuanx.wang@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>