ci: do not dump error logs in GHA containers
On error, the build logs are displayed in GHA console and logs unless
the GITHUB_WORKFLOW env variable is set.
However, containers in GHA do not automatically inherit this variable.
We could pass this variable in the container environment, but in the
end, dumping those logs is only for Travis which we don't really care
about anymore.

Let's make the linux-build.sh more generic and dump logs from Travis
yaml itself.

Fixes: b35c4b0aa2bc ("ci: add Fedora 35 container in GHA")

Signed-off-by: David Marchand <david.marchand@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>