event/cnxk: add SLMTST support to Tx adapter
Scheduled LMTST uses in-core LSW (LMTST scheduling widget) to
coordinate with SSO and send a LMTST to the destination
coprocessor without the need for the core to be the head of
the scheduling context it is currently holding.

Use SLMTST to send mbuf to NIX-TX for transmit. SLMTST only
supports transmitting a single WQE.

Signed-off-by: Pavan Nikhilesh <pbhagavatula@marvell.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>