vmxnet3: set reta size
Currently the driver assumes that the size of the reta table is
four times the number of rx queues at multiple places. However,
it sets it to four times the maximum number of queues (4 * 32 = 128)
when it first initializes the device. Change the other uses to use
the stored value, not the assumed default.

Signed-off-by: Pankaj Gupta <pagupta@vmware.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>