app/testpmd: perform SW IP checksum for GRO packets
The GRO library doesn't re-calculate checksums for
merged packets. If users want the merged packets to
have correct IP checksums, they should select HW IP
checksum calculation for the port which the merged
packets are transmitted to. But if the port doesn't
support HW IP checksum, users may perform a SW IP
checksum. This patch add the code about it.

Signed-off-by: Wenwu Ma <wenwux.ma@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>