ixgbe: make hotplug detection aware of changed SFPs
SFP hotplug may require setup/configuration of the SFP in the event that
the SFP is different from the previously installed SFP.  Ensure that the
periodic service task can handle this situation.  Teach the interrupt
path that an SFP removal means that an SFP insertion will need to be
setup again.

Signed-off-by: Jeff Daly <jeffd@silicom-usa.com>
Inspired-by: Stephen Douthit <stephend@silicom-usa.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>